HttpSys logs and samples cleanup (#17971)
