+++ src/Servers/HttpSys/samples/SelfHostServer/Program.cs
@@ -1,0 +1,32 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Server.HttpSys;
+using Microsoft.Extensions.Hosting;
+
+namespace SelfHostServer
+{
+    public static class Program
+    {
+        public static void Main(string[] args)
+        {
+            CreateHostBuilder(args).Build().Run();
+        }
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>()
+                    .UseHttpSys(options =>
+                    {
+                        options.UrlPrefixes.Add("http://localhost:5000");
+                        // This is a pre-configured IIS express port. See the PackageTags in the csproj.
+                        options.UrlPrefixes.Add("https://localhost:44319");
+                        options.Authentication.Schemes = AuthenticationSchemes.None;
+                        options.Authentication.AllowAnonymous = true;
+                    });
+                });
+    }
+}
